Refactor: リダイレクトURLホワイトリスト検証ロジックをconcernに抽出
CustomConfirmationsControllerに実装されていたオープンリダイレクト対策
（ホスト検証・クエリパラメータ付与）を、パスワードリセット機能でも
同じロジックが必要になるためRedirectUrlWhitelistableとして切り出した。
挙動は変更していない（既存specは無修正のまま通過）。
